# docs/swagger.yaml
definitions:
  request.NewToken:
    properties:
      name:
        type: string
    type: object
  request.PushMailRequest:
    properties:
      content:
        type: string
      sender_name:
        type: string
      subject:
        type: string
      to_mail_list:
        items:
          type: string
        type: array
      type:
        type: string
    type: object
  response.TokenDetail:
    properties:
      name:
        type: string
      token:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/mail-types:
    get:
      description: 获取邮件类型
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      summary: 获取邮件类型
      tags:
      - mail
  /api/v1/mails:
    post:
      description: 发送邮件
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: JSON数据
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.PushMailRequest'
      summary: 发送邮件
      tags:
      - Mail
  /api/v1/sign-in:
    put:
      description: 登录
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      summary: 登录
      tags:
      - Auth
  /api/v1/sign-up:
    post:
      description: 注册
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      summary: 注册
      tags:
      - Auth
  /api/v1/tokens:
    post:
      description: 生成Token
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      - description: JSON数据
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.NewToken'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.TokenDetail'
      summary: 生成Token
      tags:
      - Token
  /api/v1/wecom-text-messages:
    post:
      description: 推送企业微信机器人消息
      parameters:
      - description: Bearer 31a165baebe6dec616b1f8f3207b4273
        in: header
        name: Authorization
        required: true
        type: string
      summary: 推送企业微信机器人消息
      tags:
      - WeCom
swagger: "2.0"